Nemesis Deck View Implementation
While on one of 'You Win' or 'Game Over' screens, a 'View Nemesis Deck'
button is added that allows the player to view their opponent's full
deck. This is accomplished with a call the base game's
`G.UIDEF.view_deck` function with a tight, temporary replacement of `G.playing_cards` with a
newly added `MP.GAME.nemesis_cards`.

With regards to network action handlers, this feature is implemented in
a similar fashion to how the end game jokers are received.