Fix Crash on 'View Nemesis Deck' button press in reused lobby
In this commit a crash is fixed when the 'View Nemesis Deck' button is
pressed after a lobby is reused for a second match and beyond.

The `MP.GAME.nemesis_cards` has also been renamed to `MP.nemesis_cards`
to better fall in line with the structure of the `MP` object.

Currently, when the user hits 'Back' on the 'View Nemesis Deck' menu,
another set of `getEndGameJokers` and `receiveEndGameJokers` requests
are being processed by the server. This can be fixed by caching the initial
`receiveEndGameJokers` request and only doing this once up front.
I would have implemented this now, but this change fell outside of the
scope of this Pull Request. If permissible, I would like to change this
as well.