Updated keyboard offsets for controller mp menus
Steammodded appears to have a bug that makes certain menu elements overlap the controller keyboard, making it look bad, and also harder to use. This changes the offsets of all the menus specific to the balatro multiplayer mod, to not be centered, which helps avoid this issue until it's fixed. It's a bit hacky, but it gets the job done and works on other screen sizes (as far as I can tell in testing on windowed mode)